further modify code generation monad
This is now split into a monad stack. The basic code generation state for expressions is much the same, but now runs over an outer monad that keeps track of external declarations that need to be added when defining the module.

Also update Arithmetic.hs so that the standard C library functions are called instead of the llvm.* intrinsics. Add function attributes ReadNone and NoUnwind to these calls as well.