codegen: change block chain from map to sequence
The block chain is now just a sequence, and new instructions are always appended
to the end. This simplifies the process somewhat, and as a happy consequence the
generated code is more readable. Now, only the Phi operation requires searching
the block chain for the appropriate place to insert the instruction.