change the Target setup to take an AST module (wrapper)
We can't have the compile phase generate a C++ version of a module to keep around and then pass to the target compilation phases. This object gets finalized as soon as the continuation exits, so attempting to return and use the module afterwards will use undefined memory.

We may be able to get around this later, but for now the compilation targets are given just the basic AST.Module.