implement parallel reductions to a single element
This is a two-step process. Threads individually reduce a non-empty chunk and store the result into a temporary array. The master thread then reduces this array together with the seed element (if applicable) to produce the final result.

Multidimensional reductions and foldAll when executed in a single thread still only require one step.