add extensible implementations of memory table and nursery
These are parameterised by the type of the remote memory container (c :: * -> *).

Compared to the implementation used in the accelerate-cuda backend, we are using a HashMap instead of a HashTable for the underlying storage mechanism. Additionally, the HostArray does not wrap the array data into a StablePtr, but instead uses the raw pointer value. This is fine, as the underlying ST array is pinned so that it can be accessed from foreign code, and thus will not be moved by the Haskell GC, which would change the pointer value.