change Remote interface
This is simplified to the core memory transfer operations for dealing with backends: copying to/from the host and direct copy between peers. This moves the function indexArray out of class: we don't have to do a cross-device index.

Also generalise the interface to Arrays, and provide a new helper function for traversing the Arrays structure.