add a verify step after lowering the LLVM AST to C++ objects
This is done in debug mode only.

Note that if LLVM is compiled with assertions, these same checks may trigger an assertion before this function is reached  (e.g. type mismatch). However, if we do make it this far, we get an analysis of the entire module structure (e.g. blocks without terminators).