fix possible reduceWarpTree race
In the first step of a full warp reduce, do a full __syncthreads() instead of just a __threadfence_block(). Since this case processes [0,2*warpSize) elements, must ensure that the upper warp has reached the same point in the program. Otherwise, if it is far behind, it may not yet have issued the store, and thus can not be synchronised against with just a memory fence.