a framework for vertically compose-able resource stacks
A resources consists of a startup and work search function, both implemented as callbacks. When resources are combined, these continuations are executed in a left-biased manner. For the work search, this means that if the first component always returns Just work, then the second resource is never searched.

The core of the Worker state is a 1.5-ended deque containing the range of  elements left to process. Threads use this to steal work from each other.