rename nvvm -> ptx
This naming is a bit more accurate as to what is being generated, since both the LLVM NVPTX backend and libNVVM both output PTX, which is then fed to the CUDA driver to be compiled to an executable. Thus, Accelerate is converted to LLVM IR and thus to PTX before being assembled by the driver.

But, *ugh* ...