mark architecture-parameterised functions as INLINEABLE
This will put a copy of the function into the interface file, so that later we (or better yet, ghc) can specialise the type at the use site. This should avoid dictionary lookups in the client backends, where the architecture parameter is always going to be fixed.

Must test whether an accompanying SPECIALISE pragma is also required at the use site.