Redo the Async class
The Async class now has associated types for both Events and Streams (i.e. threads). Class functions are now simpler and focus on creating raw streams and using events to synchronise them. This is possible now that Streams and Events are garbage collected, so there is no need to provide combinators that bracket creation/destruction of these things.