Only copy results back to the host as they are demanded
The intent is that if we have some computation `Acc (a,b)` but we only actually require the `a` part, then `b` should not be copied back to the host at all. For example, in a simulation `b` could represent the (potentially very large) current state of the simulation, and `a` some test to determine whether or not the simulation has converged yet. This change allows us to use `a` on the host to decide what to do next, while leaving `b` on the device for later simulation steps without round-tripping to the host.

There main disadvantage of this is probably that we can no longer use multiple memcpy engines. The arrays are still transferred in non-default streams however, so that is still possible, just not automatic.