ptx: don't try to allocate size zero arrays
CUDA will barf with an invalid argument error if we attempt this. Instead, just return a nullPtr, to indicate success to the allocator without consuming any device memory. This should be safe as no empty arrays should every be dereferenced...

We should consider whether or not to make the underlying allocator fast-path zero sized arrays, as they are still consuming entries in the cache tables.