Add backend-specific metadata to generated kernels
In particular, this is used by the PTX backend to attach the launch configuration for that kernel. This places the requirements for the kernel next to the generated code, rather than off in a different module, where it is easy for the two to get out of sync.

Moreover, it allows multi-kernel modules to provide a different configuration to each kernel. For example, multidimensional exclusive reductions where the innermost dimension is empty fill the outer dimensions with the initial element: this is implemented as a different generate kernel, which has a different launch configuration and resource usage profile than the main reduction kernel.