Add helpers for writing lock-free permute update functions
These are based on the compare-and-swap instruction and can be used to implement operations such as atomic-min or atomic-add at types that are not supported by the native instruction.

This also allows us to relax the restriction for when the RMW version should be emitted to any scalar type, not just integral types. The NVPTX backend in particular has native support for atomic-add/sub for floating point types (depending on hardware generation), so this change allows us to exploit that.