ptx: implement segmented reductions
Note that this is a bit strange because we need to work around some CUDA oddity / possible NVPTX bug. The inner per-segment loop must be sure to keep all threads of the block alive even if they do not contribute anything for this segment, otherwise they will exit and contribute nothing further to the computation of _any_ segment. Introduction and use of the 'undef' constant value allows us to work around the issue.