More explicitly handling of execution streams
Now we can both create a new concurrent execution stream, as well as say that no more work will be submitted to the given stream (new).

The motivation is that in the ptx backend, while execution streams are garbage collected and will eventually be scavenged, this happens far too infrequently. The lulesh program would create thousands of streams between GC cycles and could often run out of device memory because of this.

Also see note: [Finalising execution streams].