ptx: utilities for declaring dynamically scheduled work queues
Instead of thread blocks being statically asigned which items to execute (e.g. by striding the input array by the grid dimension) this provides a simple abstraction for a dynamically managed queue of items. This is particularly useful for unbalanced workloads, or when we need some control on the order the input is processed in.

This is implemented by incrementing an atomic counter and sharing the result to all threads in the block via shared memory. As such, dequeing items from the queue requires an additional round-trip to global memory + synchronisation (~1000 cycles).