Reorganise thread gang & scheduler
The crux is that we now require exclusive access to the thread gang when executing operations (it is wrapped in an MVar). This means that we can safely call run* from multiple threads concurrently given the same execution target. Exclusivity is on a per-operation level (e.g. map, generate), not per each call to run.

The thread gang is also no longer retained as part of the Native execution structure, which was incorrect.

Fixes: AccelerateHS/accelerate#336