sequences: chunk scheduler doesn't actually take into account parallelism
The chunk size scheduler doesn't actually take into account parallelism, and is making decisions based on the wall-clock time only. If you have multiple cores (i.e. CPU backend) it does not consider the overall productivity.

It also has the problem where it can get stuck in a death spiral: decreasing the chunk size causes the program to perform worse so the chunk size is reduced again further degrading performance, ad infinitum. I think that this does not take into account the possibility that decreasing chunk sizes decreases the amount of available parallelism; e.g. if you have multiple cores; or when decreasing towards the constant factor overhead of launching a computation.

Change the metric that the CPU backend uses to speedup factor (wall time / cpu time). This makes it somewhat less likely to get stuck in the death spiral, but not entirely.