Add Mach-O object file parser and loader
Why do we need to do this, when GHC already includes a runtime linker? The primary reason is that GHC's linker does not respect the load commands of the object file, which describe how the sections of the object should be laid out in memory. For the most part one can get away with this, but it is problematic for us, as we generate AVX instructions which require particular alignments.

Luckily, the kinds of objects that we produce should also be very simple; for example, containing no external symbols which need to be resolved. Thus, loading the objects ourself should not be too burdensome a task.