native: execute via dynamically linked object files
Insteading of lowering through MCJIT, we output an object file which we parse and load into executable pages. The advantage of this is that (a) we don't need to have threads around preventing the MCJIT bracketed functions from executing; and (b) paves the way for on-disk caching and compile-time linking.