Add initial version of compiler plugin
This compiler plugin extracts annotation pragmas from the source program which specify [the path to] extra object files to be linked in with the final executable. The intent is that we will add these annotations automatically as part of an ahead-of-time compilation step.

This is a bit of a hack, but seems to work. It does require a different method depending on whether we are compiling normally or in interactive mode; a small wrapper program is provided for the latter which simply passes in the interactive mode flag automatically.