merge 'acelerate-llvm-plugin' into 'accelerate-llvm-native'
The CPU backend is the only package which requires the plugin architecture, so might as well merge the two. There is an argument for keeping them separate, however, since we can't for example load the plugin into GHCi and use it at the same time, which is why the plugin modules must live in a different subdirectory to the main backend modules.