pass the identifier as an argument to the kernel code generator functions
Rather than storing this in the code generator state structure, explicitly pass it to each of the code generation functions. Previously, looking up the UID in order to generate the kernel function names (for the native backend) would force code generation to run, but this is unnecessary if the object code is cache on disk or already loaded into memory. This change prevents that from happening.