native: add custom Setup.hs infrastructure
This provides a custom cabal 'buildHook' for use when building libraries which use 'runQ'. This takes the most unfortunate path and duplicates a lot of code from the Cabal library itself, as I could not yet find a reliable way around this, but in the future we should fix this.