ptx: have 'run' execute on any available GPUs
Introduces a shared pool of the available GPUs in the system. Each time 'run' is called it acquires the next available GPU to execute the computation (or blocks until ones is available). The idea being that we can call 'run' from multiple threads and the work will be spread out over as many GPUs as there are available.