Use explicit futures for representing asynchronous computations
This is a work-in-progress overhall of the execution phase, but currently works end-to-end for both backends with simple kernels. In particular, the `Async` class now has an associated type representing the monad in which computations are executed; in the PTX backend this is a reader monad, whereas the Native backed has a contination+reader stack.