Fix AccelerateHS/accelerate#423
In the false branch, the current thread is out-of-bounds and has no further value to read. However, rather than just returning an old value (it is ignored in later steps, so its actual value is not important) we must explicitly return the special 'undef' value. Possibly an NVPTX bug incorrectly trying to remove branches / improve memory cealescing?