Some porting progress (...)
The code seems to explicitly load libdevice.bc and _modify_ the module,
changing some names to internal, before using it in linking with the
rest of the program. This is tricky because with an IR pretty-printing
pipeline we don't want to load modules in our process and do stuff with
them.

Not sure how to handle this.