Fix hang in scanl1, move syncthreads to be under uniform control flow and document uniformity of control flow
For a scanl1, we performed 'scanBlock' under non-uniform control flow. This is forbidden, as that function will run __syncthreads. This commit fixes that.

Furthermore, there was another __syncthreads just after a non-uniform conditional. This commit moves that __syncthreads to before the conditional, such that it is under uniform control flow.

Finally, I added comments to explain why key parts of the scans are under uniform control flow