ptx: Move nanosleep inline asm to a hand-written function
llvm-pretty does not (yet) support inline assembly in call instructions,
so we manually define a separate function the final LLVM IR module that
does the requisite magic, and call that function from the place where we
want the nanosleep instruction. Because the hand-written function is
marked 'alwaysinline' (and because it's tiny anyway), LLVM will
immediately inline it and no performance is lost.