Vendor in llvm-pretty using cabal mixins
We are using a fork and want the user of accelerate-llvm to not have to
pull in a git dependency for llvm-pretty. Thus we vendor it in as a git
submodule. We ensure the code is available to all three libraries
(without compiling it in thrice) in this package
(accelerate-llvm{,-native,-ptx}) by exposing renamed versions from
accelerate-llvm; we replace the Text.LLVM prefix with
Data.Array.Accelerate.LLVM.Internal.LLVMPretty. These modules
unfortunately have to be public so that the -native and -ptx libraries
can see them.