Remove (currently broken anyway) nvvm support
NVVM requires either LLVM 7 (for which accelerate codegen has already
been broken for quite a while) or LLVM 20 + a Blackwell (CC >= 10) GPU.
We don't have a Blackwell GPU, so we cannot easily test LLVM 20 NVVM
even if we'd reintroduce the functionality.

As half the NVVM functionality got lost in the llvm-pretty port anyway,
let's remove it entirely.