ptx: Fix synchronisation of stencil border kernels
The border kernels did synchronise with the parent stream at the _end_
of the border kernel (so that the Future that contains the result of the
stencil operation is dependent on the borders too), but not at the
_start_ of the border kernel (meaning that the borders can start
executing in parallel with the computation of the argument of the
stencil operation). The latter is clearly wrong, and this commit fixes
it. This fixes a stencil nondeterminism bug that we were having
(nondet-stencil in https://github.com/tomsmeding/accelerate-tests).